Introduced role-based visibility for admin users
My goal is to make certain features accessible as admins and not as normal users, so I added an endpoint /role so I can see whether the system can recognise admins vs users.

userRole.js serves as a utility function to fetch the role. The frontend js(in public/scripts) needs to fetch the user’s role and use it to control UI visibility. Defaults to 'user' if there is an error